requirements.txt
asyncio
azure-common
azure-core
azure-search-documents
azure-storage-blob
beautifulsoup4
blis
catalogue
certifi
cffi
charset-normalizer
cleanco
click
colorama
cryptography
cymem
et-xmlfile
filelock
idna
isodate
Jinja2
joblib
langcodes
MarkupSafe
msrest
murmurhash
nameparser
numpy
oauthlib
openpyxl
packaging
pandas
pathy
preshed
pycparser
pydantic
pyparsing
python-dateutil
pytz
regex
requests
requests-file
requests-oauthlib
six
smart-open
soupsieve
srsly
thinc
tldextract
tqdm
typer
typing_extensions
urllib3
wasabi
XlsxWriter
ethnicolr
requests
duckdb
pandasql
pytest